64-bit changes: casts, and ioctl cmds are u_longs
SCCS-vsn: 8.4

+++ sys/net/route.h
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)route.h	8.3.1.1 (Berkeley) 12/02/94
+ *	@(#)route.h	8.4 (Berkeley) 01/09/95
  */
 
 /*
@@ -259,7 +259,7 @@ struct rtentry *
 	 rtalloc1 __P((struct sockaddr *, int));
 void	 rtfree __P((struct rtentry *));
 int	 rtinit __P((struct ifaddr *, int, int));
-int	 rtioctl __P((int, caddr_t, struct proc *));
+int	 rtioctl __P((u_long, caddr_t, struct proc *));
 int	 rtredirect __P((struct sockaddr *, struct sockaddr *,
 	    struct sockaddr *, int, struct sockaddr *, struct rtentry **));
 int	 rtrequest __P((int, struct sockaddr *,
